# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - in: query
          name: managerId
          schema: { type: ["integer", "null"] }
        - in: query
          name: status
          schema: { type: ["string", "null"] }
        - in: query
          name: tariff
          schema: { type: ["string", "null"] }
        - in: query
          name: month
          schema: { type: ["string", "null"] }
        - in: query
          name: search
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /leads/{id}/move:
    patch:
      operationId: moveLead
      tags: [leads]
      summary: Move lead to a new status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveLeadInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
  /leads/{id}/notes:
    post:
      operationId: addLeadNote
      tags: [leads]
      summary: Add activity note to lead
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityNote"
        "404":
          description: Not found
  /leads/{id}/activities:
    get:
      operationId: listLeadActivities
      tags: [leads]
      summary: Get activity history for a lead
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityNote"
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - in: query
          name: managerId
          schema: { type: ["integer", "null"] }
        - in: query
          name: tariff
          schema: { type: ["string", "null"] }
        - in: query
          name: paymentMethod
          schema: { type: ["string", "null"] }
        - in: query
          name: month
          schema: { type: ["string", "null"] }
        - in: query
          name: status
          schema: { type: ["string", "null"] }
        - in: query
          name: search
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete payment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, clientName, status, managerId, createdAt]
      properties:
        id: { type: integer }
        clientName: { type: string }
        phone: { type: ["string", "null"] }
        telegram: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        product: { type: ["string", "null"] }
        tariff: { type: ["string", "null"] }
        price: { type: ["number", "null"] }
        netProfit: { type: ["number", "null"] }
        source: { type: ["string", "null"] }
        income: { type: ["string", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        managerId: { type: integer }
        managerName: { type: ["string", "null"] }
        paymentDate: { type: ["string", "null"] }
        paymentType: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    LeadInput:
      type: object
      required: [clientName, status, managerId]
      properties:
        clientName: { type: string }
        phone: { type: string }
        telegram: { type: string }
        email: { type: string }
        product: { type: string }
        tariff: { type: string }
        price: { type: number }
        netProfit: { type: number }
        source: { type: string }
        income: { type: string }
        status: { type: string }
        notes: { type: string }
        managerId: { type: integer }
        paymentDate: { type: string }
        paymentType: { type: string }
    LeadUpdate:
      type: object
      properties:
        clientName: { type: string }
        phone: { type: string }
        telegram: { type: string }
        email: { type: string }
        product: { type: string }
        tariff: { type: string }
        price: { type: number }
        netProfit: { type: number }
        source: { type: string }
        income: { type: string }
        status: { type: string }
        notes: { type: string }
        managerId: { type: integer }
        paymentDate: { type: string }
        paymentType: { type: string }
    MoveLeadInput:
      type: object
      required: [status]
      properties:
        status: { type: string }
    ActivityNote:
      type: object
      required: [id, leadId, content, createdAt]
      properties:
        id: { type: integer }
        leadId: { type: integer }
        content: { type: string }
        authorName: { type: ["string", "null"] }
        createdAt: { type: string }
    ActivityNoteInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
    Payment:
      type: object
      required: [id, clientName, tariff, revenue, paymentDate, managerId]
      properties:
        id: { type: integer }
        orderNumber: { type: ["integer", "null"] }
        clientName: { type: string }
        telegram: { type: ["string", "null"] }
        tariff: { type: string }
        revenue: { type: number }
        netProfit: { type: ["number", "null"] }
        receivable: { type: ["number", "null"] }
        paymentMethod: { type: ["string", "null"] }
        paymentDate: { type: string }
        managerId: { type: integer }
        managerName: { type: ["string", "null"] }
        paymentSchedule: { type: ["string", "null"] }
        status: { type: ["string", "null"] }
        createdAt: { type: string }
    PaymentInput:
      type: object
      required: [clientName, tariff, revenue, paymentDate, managerId]
      properties:
        orderNumber: { type: integer }
        clientName: { type: string }
        telegram: { type: string }
        tariff: { type: string }
        revenue: { type: number }
        netProfit: { type: number }
        receivable: { type: number }
        paymentMethod: { type: string }
        paymentDate: { type: string }
        managerId: { type: integer }
        paymentSchedule: { type: string }
        status: { type: string }
    PaymentUpdate:
      type: object
      properties:
        clientName: { type: string }
        telegram: { type: string }
        tariff: { type: string }
        revenue: { type: number }
        netProfit: { type: number }
        receivable: { type: number }
        paymentMethod: { type: string }
        paymentDate: { type: string }
        managerId: { type: integer }
        paymentSchedule: { type: string }
        status: { type: string }
...