# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadUpdate {
  clientName?: string;
  phone?: string;
  telegram?: string;
  email?: string;
  product?: string;
  tariff?: string;
  price?: number;
  netProfit?: number;
  source?: string;
  income?: string;
  status?: string;
  notes?: string;
  managerId?: number;
  paymentDate?: string;
  paymentType?: string;
}